backend_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
stripe==7.8.0
authlib==1.2.1
itsdangerous==2.1.2
jinja2==3.1.2
aiosmtplib==3.0.1
websockets==12.0
email-validator==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
